requirements.txt
python==3.9.16
pytorch==1.12.1
pandas==1.5.3
numpy==1.24.3
tqdm==4.65.0
matplotlib==3.7.1
pygeohash==1.2.0
torch_geometric==2.3.1
pyarrow==11.0.0
transformers==4.29.2